% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_delta.R
\name{check_delta}
\alias{check_delta}
\title{check the delta max of diff(A) for the leaf data}
\usage{
check_delta(leaf_df, delta_max = 0.2)
}
\arguments{
\item{delta_max}{the filter to discard unstable values of racir data}

\item{empty_df}{the measurement data frame of the chamber with leaf}
}
\description{
check the delta max of diff(A) for the leaf data
}
