% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_delta_empty.R
\name{check_delta_empty}
\alias{check_delta_empty}
\title{check the delta max of diff(A) for the empty chamber data}
\usage{
check_delta_empty(empty_df, delta_max = 0.05)
}
\arguments{
\item{empty_df}{the measurement data frame of empty chamber}

\item{delta_max}{the filter to discard unstable values of racir data}
}
\description{
check the delta max of diff(A) for the empty chamber data
}
