% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_racir.R
\name{compile_racir}
\alias{compile_racir}
\title{Compiles .csv file outputs from {racircalbatch}. Will also compile .csv data generally, as long as the headers are the same.}
\usage{
compile_racir(outputfile)
}
\arguments{
\item{outputfile}{Specifies the name for the output file, e.g. "output.csv"}
}
\value{
Compiles .csv files for further analysis
}
\description{
\code{compile_racir} Compiles .csv files into a file named "Compiled_Data.csv"
}
