% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_leaf.R
\name{correct_leaf}
\alias{correct_leaf}
\title{Corrects rapid A/Ci response (RACiR) data from leaves using empty chamber data.}
\usage{
correct_leaf(df_empty, df_leaf)
}
\arguments{
\item{df_empty}{empty chamber racir data(for correcting the measured leaf data)}

\item{df_leaf}{racir data with leaf}
}
\description{
Corrects rapid A/Ci response (RACiR) data from leaves using empty chamber data.
}
