% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cut.R
\name{find_cut}
\alias{find_cut}
\title{find the outliers of the measured CO2}
\usage{
find_cut(df)
}
\arguments{
\item{df}{Calibration (empty chamber) rapid A/Ci response file}
}
\description{
find the outliers of the measured CO2
}
