% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racircalcheck_advanced.R
\name{racircalcheck_advanced}
\alias{racircalcheck_advanced}
\title{Allows visual checking of rapid A/Ci response (RACiR) calibration data using empty chamber data.}
\usage{
racircalcheck_advanced(calfile, mincut, maxcut, skiplines, filetype,
  digits)
}
\arguments{
\item{calfile}{Calibration (empty chamber) rapid A/Ci response file}

\item{mincut}{Minimum cutoff value for reference CO2 (CO2_r). Used to cut out the data from the initial chamber mixing. Default value is set to the minimum COR_r value.}

\item{maxcut}{Maximum cutoff value for reference CO2 (CO2_r). Used to cut out the data from the end of the response. Not needed in all cases. Default value is set to the maximum COR_r value.}

\item{skiplines}{A number specifying the number of header lines to skip.}

\item{filetype}{Specifies input datatype. Defaults to 6800.}

\item{digits}{Specifies rounding for groups. Defaults to -2 (100s).}
}
\value{
racircal returns a data frame with corrected RACiR data
}
\description{
\code{racircalcheck_advanced} Used to check range of calibration file. Produces diagnostic graphs of A vs. Ci for quality control. Output includes a data frame in the global environment "corrected_data" and a csv file of the form "datafile.csv".
}
