% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_6800.R
\name{read_6800}
\alias{read_6800}
\title{Reads files from the Li-Cor 6800}
\usage{
read_6800(filename, skiplines)
}
\arguments{
\item{filename}{A character string of the form: "mydata".}

\item{skiplines}{A number specifying the number of header lines to skip.}
}
\value{
read_6800 imports a Li-Cor 6800 file as a data frame
}
\description{
\code{read_6800} Reads Li-Cor 6800 files, which are delimited by spaces and tabs.
}
