% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Walz.R
\name{read_Walz}
\alias{read_Walz}
\title{Reads files from the Walz system}
\usage{
read_Walz(filename, skiplines)
}
\arguments{
\item{filename}{A character string of the form: "mydata".}

\item{skiplines}{A number specifying the number of header lines to skip.}
}
\value{
read_Walz imports a Walz file as a data frame
}
\description{
\code{read_Walz} Reads Walz files, which are delimited by commas.
}
