% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data_empty.R
\name{tidy_data_empty}
\alias{tidy_data_empty}
\title{Tidy the racir data by use the delta of diff for A}
\usage{
tidy_data_empty(
  df,
  delta_max = 0.2,
  upper_A = 2,
  lower_A = -2,
  reverse = FALSE
)
}
\arguments{
\item{df}{the dataframe of the racir data}

\item{delta_max}{the filter to discard unstable values of racir data}

\item{upper_A}{use to filter ucorrect A value (upper limit)}

\item{lower_A}{use to filter ucorrect A value (lower limit)}

\item{reverse}{whether the co2R is from low to high, or else. default is
low to high}
}
\description{
Tidy the racir data by use the delta of diff for A
}
