% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racircal.R
\name{racircal}
\alias{racircal}
\title{Corrects rapid A/Ci response (RACiR) data from leaves using empty chamber data.}
\usage{
racircal(calfile, mincut, maxcut, datafile, skiplines)
}
\arguments{
\item{calfile}{Calibration (empty chamber) rapid A/Ci response file}

\item{mincut}{Minimum cutoff value for reference CO2 (CO2_r). Used to cut out the data from the initial chamber mixing. Default value is set to the minimum COR_r value.}

\item{maxcut}{Maximum cutoff value for reference CO2 (CO2_r). Used to cut out the data from the end of the response. Not needed in all cases. Default value is set to the maximum COR_r value.}

\item{datafile}{Name of the data file to be corrected}

\item{skiplines}{A number specifying the number of header lines to skip.}
}
\value{
racircal returns a data frame with corrected RACiR data
}
\description{
\code{racircal} Corrects your RACiR data files based on a calibration file. Produces diagnostic graphs of A vs. Ci for quality control. Output includes a data frame in the global environment "corrected_data" and a csv file of the form "datafile.csv".
}
