% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racircal_advanced.R
\name{racircal_advanced}
\alias{racircal_advanced}
\title{Corrects rapid A/Ci response (RACiR) data from leaves using empty chamber data.}
\usage{
racircal_advanced(calfile, mincut, maxcut, datafile, skiplines, fileID,
  filetype, digits)
}
\arguments{
\item{calfile}{Calibration (empty chamber) rapid A/Ci response file}

\item{mincut}{Minimum cutoff value for reference CO2 (CO2_r). Used to cut out the data from the initial chamber mixing. Default value is set to the minimum COR_r value.}

\item{maxcut}{Maximum cutoff value for reference CO2 (CO2_r). Used to cut out the data from the end of the response. Not needed in all cases. Default value is set to the maximum COR_r value.}

\item{datafile}{Name of the data file to be corrected.}

\item{skiplines}{A number specifying the number of header lines to skip.}

\item{fileID}{Specifies output file name when using dataframes}

\item{filetype}{Specifies input datatype. Defaults to 6800.}

\item{digits}{Specifies rounding for groups. Defaults to -2 (100s).}
}
\value{
racircal_advanced returns a data frame with corrected RACiR data
}
\description{
\code{racircal_advanced} Experimental correction for RACiR data.
}
