% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy the racir data by use the delta of diff for A}
\usage{
tidy_data(df, delta_max = 0.2)
}
\arguments{
\item{df}{the dataframe of the racir data}

\item{delta_max}{the filter to discard unstable values of racir data}
}
\description{
Tidy the racir data by use the delta of diff for A
}
